+++ ft-session-replay/src/main/java/com/ft/sdk/sessionreplay/utils/Utils.java
@@ -2,25 +2,25 @@ package com.ft.sdk.sessionreplay.utils;
 
 public class Utils {
 
-    public static long densityNormalized(long size, float density) {
+    public static double densityNormalized(long size, float density) {
         if (density == 0f) {
             return size;
         }
-        return (long) (size / density);
+        return size / density;
     }
 
-    public static long densityNormalized(float size, float density) {
+    public static double densityNormalized(float size, float density) {
         if (density == 0f) {
-            return (long) size;
+            return size;
         }
-        return (long) (size / density);
+        return size / density;
     }
 
-    public static int densityNormalized(int value, float density) {
+    public static double densityNormalized(int value, float density) {
         if (density == 0f) {
             return value;
         }
-        return (int) (value / density);
+        return value / density;
     }
 
     public static CharSequence[] convertIntArrayToCharSequenceArray(int[] intArray) {
